requirements.txt
asgiref==3.11.1
certifi==2026.1.4
charset-normalizer==3.4.4
distlib==0.4.0
Django==6.0.2
django-ckeditor==6.7.3
django-cors-headers==4.9.0
django-jazzmin==3.0.2
django-js-asset==3.1.2
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.20.3
gunicorn==25.0.3
idna==3.11
packaging==26.0
pillow==12.1.0
platformdirs==4.5.1
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dotenv==1.2.1
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.3
virtualenv==20.36.1
whitenoise==6.11.0
